make PluginConfig and dependent types public
