+++ packages/core-api/src/plugin/Plugin.tsx
@@ -14,54 +14,18 @@
  * limitations under the License.
  */
 
-import { ComponentType } from 'react';
 import {
+  PluginConfig,
   PluginOutput,
-  RoutePath,
-  RouteOptions,
-  FeatureFlagName,
   BackstagePlugin,
 } from './types';
 import { validateBrowserCompat, validateFlagName } from '../app/FeatureFlags';
-import { RouteRef } from '../routing';
 import { AnyApiFactory } from '../apis';
-
-export type PluginConfig = {
-  id: string;
-  apis?: Iterable<AnyApiFactory>;
-  register?(hooks: PluginHooks): void;
-};
-
-export type PluginHooks = {
-  router: RouterHooks;
-  featureFlags: FeatureFlagsHooks;
-};
-
-export type RouterHooks = {
-  addRoute(
-    target: RouteRef,
-    Component: ComponentType<any>,
-    options?: RouteOptions,
-  ): void;
-
-  /**
-   * @deprecated See the `addRoute` method
-   */
-  registerRoute(
-    path: RoutePath,
-    Component: ComponentType<any>,
-    options?: RouteOptions,
-  ): void;
-};
-
-export type FeatureFlagsHooks = {
-  register(name: FeatureFlagName): void;
-};
 
 export class PluginImpl {
   private storedOutput?: PluginOutput[];
 
-  constructor(private readonly config: PluginConfig) {}
+  constructor(private readonly config: PluginConfig) { }
 
   getId(): string {
     return this.config.id;
